[Update] 아기 정보 수정
Former-commit-id: 6a373b9832c53218db3ecca10afc74b56e0430d2

+++ BackEnd/src/main/java/com/ssafy/todaktodak/domain/baby/service/BabyService.java
@@ -62,7 +62,7 @@ public class BabyService {
 
         Baby findBaby = baby.get();
         String imageUrl = null;
-        if (!file.isEmpty()) {
+        if (file != null && !file.isEmpty()) {
             String preImg = findBaby.getBabyImageUrl();
             if (preImg != null && !preImg.startsWith(this.DEFAULT_IMAGE_S3)) {
                 s3Client.deleteFile(preImg);
